% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calipeRmatching.R
\name{cm_set_number_of_threads}
\alias{cm_set_number_of_threads}
\title{Sets number of parallel threads for calipeRmatching library.}
\usage{
cm_set_number_of_threads(nr_threads)
}
\arguments{
\item{nr_threads}{the desired number of threads.}
}
\value{
Zero if the number of threads is set successfully. Otherwise, the program stops.
}
\description{
Sets the number of threads for calipeRmatching library.
calipeRmatching processes observations in a parallel way. The default number of threads used
for parallel execution is the number of logical cores on the machine. That can be overwritten by
calling \code{cm_set_number_of_threads} passing as argument the desired number of threads.
Calling the function takes effect at any point of the program, not only upon loading the library.
}
\examples{

###############################################################################
# Load package
###############################################################################
library("calipeRmatching")
cm_set_number_of_threads(1) # set number of parallel threads

}
\references{
[KPV23] Kormos, V. d. Pas, V. d. Vaart (2023): Asymptotics of Caliper Matching Estimators for Average Treatment Effects, https://arxiv.org/abs/2304.08373
}
\author{
Mate Kormos
}
