% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cm_set_number_of_threads}
\alias{cpp_cm_set_number_of_threads}
\title{Sets the number or parallel threads for calipeRmatching library.}
\usage{
cpp_cm_set_number_of_threads(nr_threads)
}
\arguments{
\item{nr_threads}{the desired number of threads.}
}
\value{
Zero if NUM_THREADS sets successfully. Otherwise, the program stops.
}
\description{
Only for internal use. Users should use `cm_set_number_of_threads()` that has the same funcionality.
Sets the global variable NUM_THREADS to `nr_threads` for parallel execution.
}
\references{
[KPV23] Kormos, V. d. Pas, V. d. Vaart (2023): Asymptotics of Caliper Matching Estimators for Average Treatment Effects, https://arxiv.org/abs/2304.08373
}
\author{
Mate Kormos
}
