% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_test_cm}
\alias{cpp_test_cm}
\title{Test suite for calipeRmatching library.}
\usage{
cpp_test_cm()
}
\value{
Zero if all tests passed successfully. Otherwise, the program stops.
}
\description{
Only for internal use. Users should use `test_cm()` that has the same funcionality.
}
\references{
[KPV23] Kormos, V. d. Pas, V. d. Vaart (2023): Asymptotics of Caliper Matching Estimators for Average Treatment Effects, https://arxiv.org/abs/2304.08373
}
\author{
Mate Kormos
}
