% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calipeRmatching.R
\name{test_cm}
\alias{test_cm}
\title{Test suite for calipeRmatching library.}
\usage{
test_cm()
}
\value{
Zero if all tests passed successfully. Otherwise, the program stops.
}
\description{
Runs the test suite of the calipeRmatching library.
}
\examples{

###############################################################################
# Load package
###############################################################################
library("calipeRmatching")

###############################################################################
# Test if calipeRmatching library works as intended
###############################################################################
test_cm()

}
\references{
[KPV23] Kormos, V. d. Pas, V. d. Vaart (2023): Asymptotics of Caliper Matching Estimators for Average Treatment Effects, https://arxiv.org/abs/2304.08373
}
\author{
Mate Kormos
}
